requirements.txt
pysvg
requests
pillow
scikit-image
